requirements.txt
faiss-cpu
langchain
openai
python-dotenv
fastapi
jinja2
python-multipart
uvicorn
tiktoken
langchain
unstructured[all-docs]
pydantic
lxml
openai
faiss-cpu
tiktoken
opencv-python
boto3